requirements.txt
fastapi
uvicorn
pydantic
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp
spotipy
yt-dlp
requests
